lichessapi: use api for sending mail
There is now an api for sending mail which means we can now do it in an
official way.

Remove LICHESS_CREDS_FILE_PATH as it was only used for this one method.
Replace with LICHESS_API_TOKEN_FILE_PATH which can be generated from the
Personal API Token page [1].

[1]: https://lichess.org/account/oauth/token