fixes #449 improved accessibility of availability screen
adding an aria-label that is linked to the current setting of the
availability toggles to present hints to blind users. Not a great
solution imo but short term the best I can provide. Long term solution
might be to replace the bootstrap toggles altogether or to improve the
implementation of these hints.